backend/requirements.txt
# Beagle Backend Dependencies
# Enterprise-grade data analysis platform

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching & Queue
redis==5.0.1
celery==5.3.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
bleach==6.1.0

# AI Integration
google-generativeai==0.3.2
anthropic==0.18.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
openpyxl==3.1.2
python-magic==0.4.27

# Statistical Analysis
scipy==1.12.0
statsmodels==0.14.1
scikit-learn==1.4.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1
kaleido==0.2.1

# Storage
boto3==1.34.34
minio==7.2.3

# Monitoring & Logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
locust==2.20.1

# Code Security
bandit==1.7.7
